pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "sqlsy"
version = "0.0.2"
description = "Fill your database Quickly with dummy data"
readme = "README.md"
authors = [{ name="shoaib wani", email="shoaibwani010@gmail.com"}]
license = {file = "LICENSE"}
keywords = ["sql", "sql dummy data", "fake data"]
requires-python = ">=3.9"
classifiers = [
	"License :: OSI Approved :: MIT License",
	"Operating System :: OS Independent",
	"Programming Language :: Python :: 3.11",
	"Topic :: Database"
]
dependencies = [
	"mysql-connector-python",
	"faker",
	"rich",
	"termcolor"
]

[project.urls]
Homepage = "https://github.com/pyloris/sqlsy"

[project.scripts]
sqlsy = "sqlsy.__main__:main"